requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
redis==4.2.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
wrapt==1.14.0
